+++ packages/Streams/src/Stream.php
@@ -26,6 +26,8 @@ use Traversable;
  */
 class Stream implements StreamInterface
 {
+    use Concerns\Buffering;
+
     /**
      * Readable modes regex
      *
@@ -501,6 +503,23 @@ class Stream implements StreamInterface
         while (!feof($resource)) {
             yield $callback($resource);
         }
+    }
+
+    /**
+     * Read from this stream in chunks of specified buffer size
+     *
+     * @param  int|null  $length  [optional] Maximum bytes to read from stream. By default, all bytes left are read.
+     * @param  int  $offset  [optional] The offset on where to start to reading from.
+     * @param  int  $bufferSize  [optional] Read buffer size of each chunk in bytes.
+     *
+     * @return iterable<string>
+     */
+    public function buffer(
+        int|null $length = null,
+        int $offset = 0,
+        int $bufferSize = BufferSizes::BUFFER_8KB
+    ): iterable {
+        return $this->bufferStream($this, $length, $offset, $bufferSize);
     }
 
     /**